Move main code to cmd
